exec: return print output and evaluated values
The goal is to return print output and evaluated values so they can be
handled (printed, checked, or discarded) by the caller.

Have ExecState mutate to store (1) last value (2) print output. Not sure
if this is the best pattern, but it lifts the println! out of
`exec_decl` so it can be handled at a higher level. Right now it's
still in `exec`, but can revisit when I do more testing work.